refactor: 멤버 상세 컴포넌트 폴더링 개선
