+++ components/members/detail/MemberDetail.tsx
@@ -16,11 +16,11 @@ import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 import CareerItem from '@/components/members/detail/CareerItem';
 import EmptyProfile from '@/components/members/detail/EmptyProfile';
 import InfoItem from '@/components/members/detail/InfoItem';
+import InterestSection from '@/components/members/detail/InterestSection';
+import MemberDetailSection from '@/components/members/detail/MemberDetailSection';
 import MemberProjectCard from '@/components/members/detail/MemberProjectCard';
+import MessageSection from '@/components/members/detail/MessageSection';
 import PartItem from '@/components/members/detail/PartItem';
-import MemberDetailSection from '@/components/members/main/MemberDetail/Section';
-import InterestSection from '@/components/members/main/MemberDetail/Section/Interest';
-import MessageSection from '@/components/members/main/MemberDetail/Section/Message';
 import { DEFAULT_DATE } from '@/components/members/upload/constants';
 import { playgroundLink } from '@/constants/links';
 import { useRunOnce } from '@/hooks/useRunOnce';
